changed location with minor changes
