+++ src/server/api/trpc.ts
@@ -1,0 +1,6 @@
+import { initTRPC } from '@trpc/server';
+
+const t = initTRPC.create();
+
+export const router = t.router;
+export const publicProcedure = t.procedure;
